test(stella-pipeline): move the assemble_user_message tests beside the function
The file-size ratchet caught this against `main` where my local run judged
against an intermediate commit of this branch: `pipeline/tests.rs` is a
grandfathered god file closed to growth, and this change had put it 10 lines
over its ceiling.

The fix is the one the guard is asking for rather than a smaller edit that
happens to fit. `assemble_user_message` moved to `pipeline/user_message.rs` in
the previous commit and its four unit tests did not follow it — they were
sitting in the shared `tests.rs` testing a function that no longer lives in
the module that file tests. They move to `user_message.rs`'s own test module,
which is where a reader looking at the function will find them, and takes
`tests.rs` back under its ceiling with room rather than by one line.

Two tests join them while the subject is in view: findings lead the message
ahead of recall and the goal, and an empty findings list adds no heading to a
message that carries recall — the narrow form of the byte-identical property
the advisory stages depend on.

No baseline was raised.

Refs #2415
